correct implementations, move each of error types into their dedicated files, add destination type as one of the parameters while throwing an error
