+++ v0/util/errors/index.js
@@ -1,8 +1,9 @@
-const { TransformationError, ApiError } = require("./error-classes");
-const { RudderBaseError } = require("./base");
+const { RudderErrorBase } = require("./base");
+const TransformationError = require("./transformation");
+const ApiError = require("./api");
 
 module.exports = {
   TransformationError,
   ApiError,
-  RudderBaseError
+  RudderErrorBase
 };
